data_query_using_natural_language/requirements.txt
openai>=1.37.1, <2.0.0
python-dotenv>=1.0.1, <2.0.0
termcolor
saspy
re